requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.0
soundfile>=0.12
numpy>=1.24
torch>=2.0
chatterbox-tts
